refactor: fix `uri.js` path
