+++ front/src/components/NavigationBar/ButtonContainer/ButtonContainer.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
-import { ButtonContainerStyle, NavImage } from './ButtonContainer.style';
+import {Link} from 'react-router-dom';
+import {ButtonContainerStyle, NavImage} from './ButtonContainer.style';
 
 import NOTICE from '../../../asset/notice.png';
 import MYPAGE from '../../../asset/mypage.png';
 import LOGOUT from '../../../asset/logout.png';
 import Button from './Button/Button';
-import APP_URI from '../../../uri';
+import APP_URI from '../../../util/uri';
 
 const ButtonContainer = () => {
   // logics 로 분리예정
